// old_codebase/20260119/server/src/types.ts
export type User = {
  id: string;
  email: string;
  balance: number;
};